+++ Utility.h
@@ -5,7 +5,7 @@
 //あったら便利そうなやつまとめ
 
 void DrawRotaGraph4(int x, int y, double rate, double anglex, double angley, double anglez, int handle, int tranflag, int x_turn_flag, int y_turn_flag);
-Color GetColorHSV(float H, float S, float V);
+Color GetColorHSV(double H, double S, double V);
 Color GamingColor(double mul = 1);
 void SmartSetDrawBlendMode(int BlendMode, int Pal);
 double Rad(double angle);
